requirements.txt
aiofiles==0.8.0
aiohttp==3.8.6
aiolimiter==1.1.0
aiopg==1.4.0
aiorun==2023.7.2
aiosignal==1.3.1
aiosqlite==0.17.0
annotated-types==0.6.0
anyio==3.7.1
appdirs==1.4.4
asgiref==3.7.2
async-timeout==4.0.3
asyncio==3.4.3
asyncpg==0.29.0
asyncpraw==7.7.1
asyncprawcore==2.3.0
attrs==23.1.0
beautifulsoup4==4.12.2
bs4==0.0.1
certifi==2023.7.22
chardet==3.0.4
charset-normalizer==2.1.1
click==8.1.7
colorama==0.4.6
databases==0.8.0
DateTime==5.3
deprecation==2.1.0
Django==4.2.7
django-bootstrap-v5==1.0.11
django-extensions==3.2.3
django-filter==23.5
djangorestframework==3.14.0
drf-flex-fields==1.0.2
eventemitter==0.2.0
fastapi==0.105.0
frozenlist==1.4.0
greenlet==3.0.2
h11==0.14.0
h2==3.2.0
hpack==3.0.0
httpcore==0.17.2
httpx==0.24.1
hyperframe==5.2.0
idna==2.10
Jinja2==3.1.2
Markdown==3.5.2
markdown-it-py==3.0.0
MarkupSafe==2.1.3
mdurl==0.1.2
multidict==6.0.4
myjdapi==1.1.7
nest-asyncio==1.5.8
numpy==1.23.4
packaging==23.2
pandas==2.0.2
platformdirs